Notify user of link tags with non-relative links
Notify user when a link tag does not have a relative `href`
(as this introduces ambiguity around whether the href refers to
a file or online URL).

Previously, this case was (erroneously) not handled. User is now
notified and the tag is included in the output as any other normal tag.

Resolution of absolute directories is still not implemented.